Adding more runescript support.
This is not currently safe.
We need a way to have an effect build it's state from scratch. As well
as keep that state around between instances, but not be sendable.

My next thoughts are to keep a global context for each script, and let
the unit be compiled _once_ per Effect. This means EffectType -> Effect
won't be one to one anymore, since I would have to construct a new
RuneScript each time EffectType::RuneScript is shared.