Works(?).
Or at least it works in-so-far as I can run a script every _duration_
that is returned by `render`.

Whether or not this actually will control lights, we'll have to see.
Otherwise, I'm very excited.

We also need to make sure we can have the scripts write directly to
the pixel slice.

Also included is some cleanup and extensions to rainbows and cleanup of
unused dependencies. Since I couldn't be bothered to clean this commit up.